fix: vars in output were not being expanded into refs
Also, ensure that hashes are correctly updated when targets are built,
to prevent a build of every input.

This successfully allows arbitrarily dependencies to build in the
correct order.

eg. In the following example I delete a generated .proto file then
rebuild the binary that depends on the .pb.go files generated from that
.proto file. This regenerates the .proto, then the .pb.go files, then
finally the binary.

    🐚 ~/dev/ftl $ rm ./protos/xyz/block/ftl/v1/schema/schema.proto
    🐚 ~/dev/ftl $ ../bit/build/bit build/release/ftl-runner
    …ma/schema.proto| $ ftl schema protobuf > protos/xyz/block/ftl/v1/schema/schema.proto
    …e/console.pb.go| $ buf format -w
    …e/console.pb.go|   buf lint
    …e/console.pb.go|   (cd protos && buf generate)
    …e/console.pb.go|   (cd backend/common/3rdparty/protos && buf generate)
    …ease/ftl-runner| $ go build -o build/release/ftl-runner ./cmd/ftl-runner